[Seeding] Area and shelter table

+++ database/migrations/2017_10_23_040652_create_areas_table.php
@@ -14,7 +14,7 @@ class CreateAreasTable extends Migration
     public function up()
     {
         Schema::create('areas', function (Blueprint $table) {
-            $table->increments('id');
+            $table->tinyInteger('id');
             $table->string('name');
             $table->timestamps();
         });
